requirements.txt
faiss-cpu
openai
python-dotenv
rich
sentence-transformers
numpy
requests

